Random generation utilities for tests
