+++ nano/core_test/fork_cache.cpp
@@ -1,4 +1,5 @@
 #include <nano/node/fork_cache.hpp>
+#include <nano/test_common/random.hpp>
 #include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
